Remove redundant `InsufficientFunds` check and properly account for amount needed on destination chain (#34)
We were not properly accounting for the payment token when checking if
funds needed to be bridged. This only applies to the case where the
payment token is also the action token. For example, if the user wants
to transfer 10 USDC with a payment max cost of 1 USDC, the
`needsBridgedFunds` needs to check that there is 11 USDC on the
destination chain instead of just 10 USDC.

We also remove the `InsufficientFunds` check because it is redundant and
seems like a subset of the `assertFundsAvailable` check.