Support Quark v2 in QuarkBuilder (#83)
Adds `nonceSecret` and `replayCount` to QuarkBuilder. For replayable
operations, the QuarkBuilder will also generate the nonce using the
`nonceSecret` and `replayCount` (see `generateNonceFromSecret`
function).

We also uncover and fix a bug where `0` is being used as the `nonce` for
an operation when a nonce is not provided for an account by the client.
Now, the code reverts if a `nonceSecret` is not given for an account. We
buff up the unit tests to assert the right values are being used for the
nonces, since nonces were being checked previously.