Add Across bridge support to the QuarkBuilder (#100)
This PR adds support for bridging assets using Across to the
QuarkBuilder. This is the first time we are using an FFI in the
QuarkBuilder, so any downstream clients using this version of the
builder will need to implement the FFI at the reserved address (`address
constant ACROSS_FFI_ADDRESS = address(0xFF1000)`).